Performance improvements and bugfixes
Added a sleep to the TrackerManager mainloop when it's paused, which reduced thread from max usage.

Changed the way a mouse listener process was started to decouple the two processes.

Added fix so that only one mousebind listener can exist at a time.

Added fix so that only one profilecreator may be spawned at a time from the tracker.